requirements.txt
flask
flask_sqlalchemy
pymysql
wtforms
flask_wtf
flask_testing
pytest
pytest-cov
mysql-connector-python-rf
